Solution

+++ app/main.py
@@ -4,7 +4,7 @@ import math
 
 class Vector:
 
-    def __init__(self, x: int, y: int) -> None:
+    def __init__(self, x: float, y: float) -> None:
         self.x = round(x, 2)
         self.y = round(y, 2)
